// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/observationAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AuditPlanAC } from './auditPlanAC';
import { BaseModelAC } from './baseModelAC';
import { Disposition } from './disposition';
import { ObservationCategoryAC } from './observationCategoryAC';
import { ObservationDocumentAC } from './observationDocumentAC';
import { ObservationStatus } from './observationStatus';
import { ObservationTableAC } from './observationTableAC';
import { ObservationType } from './observationType';
import { PlanProcessMappingAC } from './planProcessMappingAC';
import { ProcessAC } from './processAC';
import { UserAC } from './userAC';

export interface ObservationAC extends BaseModelAC { 
    heading?: string;
    background?: string;
    riskAndControlId?: string;
    observations?: string;
    observationCategoryId?: string;
    isRepeatObservation: boolean;
    rootCause?: string;
    implication?: string;
    recommendation?: string;
    managementResponse?: string;
    conclusion?: string;
    personResponsible: string;
    relatedObservation?: string;
    targetDate: Date;
    linkedObservation?: string;
    processId: string;
    parentProcessId?: string;
    processName?: string;
    subProcessName?: string;
    sortOrder: number;
    statusString?: string;
    observationCategoryList?: Array<ObservationCategoryAC>;
    auditPlanList?: Array<AuditPlanAC>;
    processList?: Array<PlanProcessMappingAC>;
    parentProcessList?: Array<ProcessAC>;
    auditPlanId: string;
    personResponsibleList?: Array<UserAC>;
    userAC?: UserAC;
    observationType: ObservationType;
    disposition: Disposition;
    observationStatus: ObservationStatus;
    observationListStatus: ObservationStatus;
    observationTypeToString?: string;
    dispositionToString?: string;
    observationStatusToString?: string;
    entityId?: string;
    linkedObservationACList?: Array<ObservationAC>;
    observationTargetDate?: string;
    personResposibleName?: string;
    observationCategoryName?: string;
    auditPlanName?: string;
    isRepeated?: string;
    isDraft: boolean;
    observationFiles?: Array<Blob>;
    observationDocuments?: Array<ObservationDocumentAC>;
    observationTableList?: Array<ObservationTableAC>;
}